Move DisconnectModal components to the upper folder for reuse
