+++ js/src/settings/disconnect-modal/index.js
@@ -4,6 +4,8 @@
 import ConfirmModal from './confirm-modal';
 import './index.scss';
 
+export * from './constants';
+
 export default function DisconnectModal( props ) {
 	// TODO: add a survey modal when the method of saving the survey results has been decided.
 	return <ConfirmModal { ...props } />;
